Module aws_sdk_xray::client::fluent_builders

Utilities to ergonomically construct a request to the service.

Fluent builders are created through the Client by calling one if its operation methods. After parameters are set using the builder methods, the send method can be called to initiate the request.

Structs

BatchGetTraces

Fluent builder constructing a request to BatchGetTraces.

CreateGroup

Fluent builder constructing a request to CreateGroup.

CreateSamplingRule

Fluent builder constructing a request to CreateSamplingRule.

DeleteGroup

Fluent builder constructing a request to DeleteGroup.

DeleteSamplingRule

Fluent builder constructing a request to DeleteSamplingRule.

GetEncryptionConfig

Fluent builder constructing a request to GetEncryptionConfig.

GetGroup

Fluent builder constructing a request to GetGroup.

GetGroups

Fluent builder constructing a request to GetGroups.

GetInsight

Fluent builder constructing a request to GetInsight.

GetInsightEvents

Fluent builder constructing a request to GetInsightEvents.

GetInsightImpactGraph

Fluent builder constructing a request to GetInsightImpactGraph.

GetInsightSummaries

Fluent builder constructing a request to GetInsightSummaries.

GetSamplingRules

Fluent builder constructing a request to GetSamplingRules.

GetSamplingStatisticSummaries

Fluent builder constructing a request to GetSamplingStatisticSummaries.

GetSamplingTargets

Fluent builder constructing a request to GetSamplingTargets.

GetServiceGraph

Fluent builder constructing a request to GetServiceGraph.

GetTimeSeriesServiceStatistics

Fluent builder constructing a request to GetTimeSeriesServiceStatistics.

GetTraceGraph

Fluent builder constructing a request to GetTraceGraph.

GetTraceSummaries

Fluent builder constructing a request to GetTraceSummaries.

ListTagsForResource

Fluent builder constructing a request to ListTagsForResource.

PutEncryptionConfig

Fluent builder constructing a request to PutEncryptionConfig.

PutTelemetryRecords

Fluent builder constructing a request to PutTelemetryRecords.

PutTraceSegments

Fluent builder constructing a request to PutTraceSegments.

TagResource

Fluent builder constructing a request to TagResource.

UntagResource

Fluent builder constructing a request to UntagResource.

UpdateGroup

Fluent builder constructing a request to UpdateGroup.

UpdateSamplingRule

Fluent builder constructing a request to UpdateSamplingRule.