Module aws_sdk_xray::error

Modules

invalid_request_exception

See InvalidRequestException

resource_not_found_exception

See ResourceNotFoundException

rule_limit_exceeded_exception

See RuleLimitExceededException

throttled_exception

See ThrottledException

too_many_tags_exception

See TooManyTagsException

Structs

BatchGetTracesError

CreateGroupError

CreateSamplingRuleError

DeleteGroupError

DeleteSamplingRuleError

GetEncryptionConfigError

GetGroupError

GetGroupsError

GetInsightError

GetInsightEventsError

GetInsightImpactGraphError

GetInsightSummariesError

GetSamplingRulesError

GetSamplingStatisticSummariesError

GetSamplingTargetsError

GetServiceGraphError

GetTimeSeriesServiceStatisticsError

GetTraceGraphError

GetTraceSummariesError

InvalidRequestException

The request is missing required parameters or has invalid parameters.

ListTagsForResourceError

PutEncryptionConfigError

PutTelemetryRecordsError

PutTraceSegmentsError

ResourceNotFoundException

The resource was not found. Verify that the name or Amazon Resource Name (ARN) of the resource is

RuleLimitExceededException

You have reached the maximum number of sampling rules.

TagResourceError

ThrottledException

The request exceeds the maximum number of requests per second.

TooManyTagsException

You have exceeded the maximum number of tags you can apply to this resource.

UntagResourceError

UpdateGroupError

UpdateSamplingRuleError

Enums

BatchGetTracesErrorKind

CreateGroupErrorKind

CreateSamplingRuleErrorKind

DeleteGroupErrorKind

DeleteSamplingRuleErrorKind

GetEncryptionConfigErrorKind

GetGroupErrorKind

GetGroupsErrorKind

GetInsightErrorKind

GetInsightEventsErrorKind

GetInsightImpactGraphErrorKind

GetInsightSummariesErrorKind

GetSamplingRulesErrorKind

GetSamplingStatisticSummariesErrorKind

GetSamplingTargetsErrorKind

GetServiceGraphErrorKind

GetTimeSeriesServiceStatisticsErrorKind

GetTraceGraphErrorKind

GetTraceSummariesErrorKind

ListTagsForResourceErrorKind

PutEncryptionConfigErrorKind

PutTelemetryRecordsErrorKind

PutTraceSegmentsErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateGroupErrorKind

UpdateSamplingRuleErrorKind