Module aws_sdk_xray::model

Modules

alias

See Alias

anomalous_service

See AnomalousService

availability_zone_detail

See AvailabilityZoneDetail

backend_connection_errors

See BackendConnectionErrors

edge

See Edge

edge_statistics

See EdgeStatistics

encryption_config

See EncryptionConfig

error_root_cause

See ErrorRootCause

error_root_cause_entity

See ErrorRootCauseEntity

error_root_cause_service

See ErrorRootCauseService

error_statistics

See ErrorStatistics

fault_root_cause

See FaultRootCause

fault_root_cause_entity

See FaultRootCauseEntity

fault_root_cause_service

See FaultRootCauseService

fault_statistics

See FaultStatistics

forecast_statistics

See ForecastStatistics

group

See Group

group_summary

See GroupSummary

histogram_entry

See HistogramEntry

http

See Http

insight

See Insight

insight_event

See InsightEvent

insight_impact_graph_edge

See InsightImpactGraphEdge

insight_impact_graph_service

See InsightImpactGraphService

insight_summary

See InsightSummary

insights_configuration

See InsightsConfiguration

instance_id_detail

See InstanceIdDetail

request_impact_statistics

See RequestImpactStatistics

resource_arn_detail

See ResourceArnDetail

response_time_root_cause

See ResponseTimeRootCause

response_time_root_cause_entity

See ResponseTimeRootCauseEntity

response_time_root_cause_service

See ResponseTimeRootCauseService

root_cause_exception

See RootCauseException

sampling_rule

See SamplingRule

sampling_rule_record

See SamplingRuleRecord

sampling_rule_update

See SamplingRuleUpdate

sampling_statistic_summary

See SamplingStatisticSummary

sampling_statistics_document

See SamplingStatisticsDocument

sampling_strategy

See SamplingStrategy

sampling_target_document

See SamplingTargetDocument

segment

See Segment

service

See Service

service_id

See ServiceId

service_statistics

See ServiceStatistics

tag

See Tag

telemetry_record

See TelemetryRecord

time_series_service_statistics

See TimeSeriesServiceStatistics

trace

See Trace

trace_summary

See TraceSummary

trace_user

See TraceUser

unprocessed_statistics

See UnprocessedStatistics

unprocessed_trace_segment

See UnprocessedTraceSegment

value_with_service_ids

See ValueWithServiceIds

Structs

Alias

An alias for an edge.

AnomalousService

The service within the service graph that has anomalously high fault rates.

AvailabilityZoneDetail

A list of Availability Zones corresponding to the segments in a trace.

BackendConnectionErrors

Edge

Information about a connection between two services.

EdgeStatistics

Response statistics for an edge.

EncryptionConfig

A configuration document that specifies encryption configuration settings.

ErrorRootCause

The root cause of a trace summary error.

ErrorRootCauseEntity

A collection of segments and corresponding subsegments associated to a trace summary

ErrorRootCauseService

A collection of fields identifying the services in a trace summary error.

ErrorStatistics

Information about requests that failed with a 4xx Client Error status code.

FaultRootCause

The root cause information for a trace summary fault.

FaultRootCauseEntity

A collection of segments and corresponding subsegments associated to a trace summary

FaultRootCauseService

A collection of fields identifying the services in a trace summary fault.

FaultStatistics

Information about requests that failed with a 5xx Server Error status code.

ForecastStatistics

The predicted high and low fault count. This is used to determine if a service has

Group

Details and metadata for a group.

GroupSummary

Details for a group without metadata.

HistogramEntry

An entry in a histogram for a statistic. A histogram maps the range of observed values

Http

Information about an HTTP request.

Insight

When fault rates go outside of the expected range, X-Ray creates an insight. Insights

InsightEvent

X-Ray reevaluates insights periodically until they are resolved, and records each intermediate state in an

InsightImpactGraphEdge

The connection between two service in an insight impact graph.

InsightImpactGraphService

Information about an application that processed requests, users that made requests, or downstream services,

InsightSummary

Information that describes an insight.

InsightsConfiguration

The structure containing configurations related to insights.

InstanceIdDetail

A list of EC2 instance IDs corresponding to the segments in a trace.

RequestImpactStatistics

Statistics that describe how the incident has impacted a service.

ResourceArnDetail

A list of resources ARNs corresponding to the segments in a trace.

ResponseTimeRootCause

The root cause information for a response time warning.

ResponseTimeRootCauseEntity

A collection of segments and corresponding subsegments associated to a response time

ResponseTimeRootCauseService

A collection of fields identifying the service in a response time warning.

RootCauseException

The exception associated with a root cause.

SamplingRule

A sampling rule that services use to decide whether to instrument a request. Rule

SamplingRuleRecord

A SamplingRule and its metadata.

SamplingRuleUpdate

A document specifying changes to a sampling rule's configuration.

SamplingStatisticSummary

Aggregated request sampling data for a sampling rule across all services for a 10-second window.

SamplingStatisticsDocument

Request sampling results for a single rule from a service. Results are for the last 10

SamplingStrategy

The name and value of a sampling rule to apply to a trace summary.

SamplingTargetDocument

Temporary changes to a sampling rule configuration. To meet the global sampling target for a rule, X-Ray

Segment

A segment from a trace that has been ingested by the X-Ray service. The segment can be

Service

Information about an application that processed requests, users that made requests, or downstream services,

ServiceId

ServiceStatistics

Response statistics for a service.

Tag

A map that contains tag keys and tag values to attach to an Amazon Web Services X-Ray group or sampling

TelemetryRecord

TimeSeriesServiceStatistics

A list of TimeSeriesStatistic structures.

Trace

A collection of segment documents with matching trace IDs.

TraceSummary

Metadata generated from the segment documents in a trace.

TraceUser

Information about a user recorded in segment documents.

UnprocessedStatistics

Sampling statistics from a call to GetSamplingTargets that X-Ray

UnprocessedTraceSegment

Information about a segment that failed processing.

ValueWithServiceIds

Information about a segment annotation.

Enums

AnnotationValue

EncryptionStatus

EncryptionType

InsightCategory

InsightState

SamplingStrategyName

TimeRangeType