Module aws_sdk_xray::operation

Structs

BatchGetTraces

Retrieves a list of traces specified by ID. Each trace is a collection of segment

CreateGroup

Creates a group resource with a name and a filter expression.

CreateSamplingRule

Creates a rule to control sampling behavior for instrumented applications. Services

DeleteGroup

Deletes a group resource.

DeleteSamplingRule

Deletes a sampling rule.

GetEncryptionConfig

Retrieves the current encryption configuration for X-Ray data.

GetGroup

Retrieves group resource details.

GetGroups

Retrieves all active group details.

GetInsight

Retrieves the summary information of an insight. This includes impact to clients and

GetInsightEvents

X-Ray reevaluates insights periodically until they're resolved, and records each intermediate state as an

GetInsightImpactGraph

Retrieves a service graph structure filtered by the specified insight. The service graph is limited to only

GetInsightSummaries

Retrieves the summaries of all insights in the specified group matching the provided filter values.

GetSamplingRules

Retrieves all sampling rules.

GetSamplingStatisticSummaries

Retrieves information about recent sampling results for all sampling rules.

GetSamplingTargets

Requests a sampling quota for rules that the service is using to sample requests.

GetServiceGraph

Retrieves a document that describes services that process incoming requests, and

GetTimeSeriesServiceStatistics

Get an aggregation of service statistics defined by a specific time

GetTraceGraph

Retrieves a service graph for one or more specific trace IDs.

GetTraceSummaries

Retrieves IDs and annotations for traces available for a specified time frame using an

ListTagsForResource

Returns a list of tags that are applied to the specified Amazon Web Services X-Ray group or sampling rule.

PutEncryptionConfig

Updates the encryption configuration for X-Ray data.

PutTelemetryRecords

Used by the Amazon Web Services X-Ray daemon to upload telemetry.

PutTraceSegments

Uploads segment documents to Amazon Web Services X-Ray. The X-Ray SDK generates segment documents and sends them to the X-Ray daemon, which uploads them in

TagResource

Applies tags to an existing Amazon Web Services X-Ray group or sampling rule.

UntagResource

Removes tags from an Amazon Web Services X-Ray group or sampling rule. You cannot edit or delete system

UpdateGroup

Updates a group resource.

UpdateSamplingRule

Modifies a sampling rule's configuration.